Improve the results of synchronously executed Mutations.
When `isOneShot` in MutationOptions is set to false, multiple calls might not receive the intended results
(there was a possibility of referencing the state of the next execution).

By using a callback, the execution results of the Mutation can be awaited on the calling side,
ensuring that the results of the call are reliably referenced.