Replace `onPlaceholderData` with `onInitialData`
I have renamed `onPlaceholderData` to `onInitialData` and shifted its role to that of an initial data set. We have decided
to redesign the placeholder to be applied on a component basis in the future.

In the TanStack Query, which we reference as a benchmark, it is designed to be used differently according to the
application.

- [Initial Query Data](https://tanstack.com/query/latest/docs/framework/react/guides/initial-query-data)
- [Placeholder Query Data](https://tanstack.com/query/latest/docs/framework/react/guides/placeholder-query-data)

The `onPlaceholderData` implemented in Soil Query did not align perfectly with either, having a position somewhat in
between. From a usability perspective, it was deemed preferable to align the role more closely with Initial Query Data,
and to prepare a new implementation for the placeholder.