Introducing Combined Query
To improve the efficiency of Composable functions that define multiple keys in parallel, we have implemented a
new feature called Combined Query. This feature optimizes the recomposition process by treating multiple query keys as a
single query, but only when a composite type can be defined from these keys.

By using the Combined Query class, the number of recompositions can be significantly reduced to the same level as a
single query. This ensures that the performance of Composable functions remains high, even when working with multiple
keys simultaneously.