Optimization of Recomposition
Within the Composable functions provided by the Query library, recomposition occurs due to changes in the internal state
related to data loading. To reduce unnecessary state updates in the UI, we have implemented a RecompositionOptimizer
that omits certain state changes based on conditions.

This RecompositionOptimizer can be configured on a per-Composable function basis, and the optimization is enabled by
default. If you need to reference the omitted state items in the UI, you can disable the optimization through the
configuration settings.