Add `purgeAll` API
Currently, it is possible to explicitly delete in-memory data for queries through the `perform` function or the mutation
update process.

In #89, we added support for the Subscription API, but there is still no API available to clear in-memory data for subscriptions.
Therefore, we have implemented the `purgeAll` function, which is intended for full reset operations, such as during
sign-out processes.