Behavior Test for iOS Artifacts
The iOS target artifact generation was not functioning correctly when uploading the library to Maven Central from
CI. refs: #110

The GitHub Actions execution logs indicated warnings that the process was automatically skipped due to the execution
machine type.

```
> Configure project :soil-form
w: The following Kotlin/Native targets cannot be built on this machine and are disabled:
iosArm64, iosSimulatorArm64, iosX64
To hide this message, add 'kotlin.native.ignoreDisabledTargets=true' to the Gradle properties.
```

Based on this, the CI environment machine type needs to be changed to `macos`.
As a first step in implementing this change, the library artifacts will be output locally instead of to Maven Central,
allowing verification of correct functionality in a test job.