Implement Conditional Caching
Added a `contentCacheable` property to the key definitions for queries and subscriptions with caching mechanisms.
With this change, in addition to defining the cache retention period based on options, it's now possible to determine whether caching is enabled for each key based on the content.
This allows for usage patterns aligned with content freshness, such as "Do not cache if the list data is empty."