Implement Automatic Namespace Generation
Currently, the companion objects of `MutationId` and `SubscriptionId` have an `auto` function defined.
However, when this function is used for defining keys in a Compose runtime environment, values cannot be restored after configuration changes on the Android platform.
Therefore, we have decided to deprecate the `auto` function in the companion objects and instead provide a Compose
function with the same name.