Add Namespace Extension Property `autoCompositionKeyHash`
In #115 and #116, we introduced the `auto` extension function to assist with the automatic generation of unique namespaces for Mutations and Subscriptions.
This PR adds an alternative that allows the use of a value calculated from `currentCompositionKeyHash` instead of the
`auto` function, but only within a Composable function limited to a single key.