Tweak MutatedEffect
When it was introduced in #131, `replyUpdatedAt` was used as the default value for `keySelector`. However, an issue was discovered during unit testing.
Since `replyUpdatedAt` handles epoch time in seconds, the success callback was not triggered when updates occurred at intervals of less than a second.
The unit test, which performed multiple updates within sub-second intervals, helped uncover this issue.