Introduce idling functions for testing
Implemented `isIdleNow` function, which can be used with `IdlingResource` or `waitUntil`,and `awaitIdle` suspend
function for Compose UI testing, utilizing the `SwrCacheView` introduced in #135.

Previously, it was not possible to determine whether there were ongoing operations within the cache. With these new
functions, it is now possible to synchronize with Compose test functions for proper coordination during testing.