Simplify the Internal Lifecycle of Actor
This PR simplifies the mechanism for tracking reference counts in the Actor introduced in #44. Previously, the reference
count was managed by invoking the `launchIn(scope)` function via `XxxRef` interface, controlling the increment and decrement
of references based on the lifecycle of the CoroutineScope.

However, this approach proved inconvenient for short-lived scopes, such as those used with the prefetch function.

In this change, a UUID is assigned to each Ref instance, and a simpler attach/detach mechanism is introduced. Naturally,
Ref instances must guarantee these calls. To enforce this, the [AutoCloseable](https://kotlinlang.org/api/core/kotlin-stdlib/kotlin/-auto-closeable.html) interface is implemented, making it clear
that the close function must be invoked once the instance is no longer needed.

For short-lived use cases like prefetch, the use extension function can be employed to ensure proper cleanup reliably.

```
getQuery(key, marker).use { query -> ... }
```