Partial Integration of Await and Catch Components
We have decided to partially integrate the functionality of the `Await` and `Catch` composable functions provided in the
experimental *query-compose-runtime package*.

Previously, the implementation required placing the `Await` and `Catch` components in parallel. With this integration, a
single `Await` component can now achieve the same functionality.

Please note that the `Catch` component will still be available for independent use if needed.