Add new soil-lazyload module for better modularity
We've implemented a new standalone soil-lazyload module as part of our modularization efforts. This module provides a
completely rewritten implementation of the LazyLoad functionality that was previously contained within the
soil-query-compose-runtime module.

The LazyLoad functionality in soil-query-compose-runtime is now considered deprecated.
We plan to remove this deprecated functionality in a future release.